Moved ServiceExtenstions to its subfolder
